@@ -7,9 +7,7 @@ package com.wakamayu.jucu.health.check.injector.driver;
 
 import com.wakamayu.jucu.health.check.injector.interfaces.Driver;
 
-import javax.ejb.Stateful;
-import javax.ejb.Stateless;
-import javax.enterprise.context.RequestScoped;
+import jakarta.ejb.Stateless;
 
 import com.wakamayu.jucu.health.check.injector.configure.TracerModel;
 import com.wakamayu.jucu.health.check.injector.enums.TypeStatus;
@@ -25,8 +23,7 @@ import javax.inject.Named;
  * @author carlos
  */
 @Named("IP_PORT")
-//@Stateless
-@Stateful
+@Stateless
 public class InjectorTelnet implements Driver {
 
     @Override
