@@ -7,14 +7,14 @@ package com.wakamayu.jucu.health.check.injector.api;
 
 //import com.wakamayu.jucu.health.check.injector.abstracts.AbstractFactoryHealthCheck;
 
-
 import com.wakamayu.jucu.health.check.injector.resource.HealthResource;
 import com.wakamayu.jucu.health.check.injector.resource.MetricsPrometheus;
+
+import jakarta.annotation.PostConstruct;
 
 import java.util.HashSet;
 import java.util.Set;
 
-import javax.annotation.PostConstruct;
 import javax.enterprise.context.ApplicationScoped;
 import javax.ws.rs.ApplicationPath;
 import javax.ws.rs.core.Application;
@@ -31,11 +31,12 @@ public class ApplicationHealth extends Application {
 	 * static final Counter requests = Counter.build()
 	 * .name("requests_total").help("Total requests.").register();
 	 */
-@PostConstruct
+	@PostConstruct
 	public void init() {
-	System.out.println("PostConstruct - ApplicationHealth");
-	
-}
+		System.out.println("PostConstruct - ApplicationHealth");
+
+	}
+
 	@Override
 	public Set<Class<?>> getClasses() {
 		Set<Class<?>> classes = new HashSet<Class<?>>();
