@@ -1,9 +1,9 @@
 package com.wakamayu.jucu.health.check.injector.resource;
 
-import javax.ws.rs.Consumes;
+
 import javax.ws.rs.GET;
 import javax.ws.rs.Path;
-import javax.ws.rs.Produces;
+
 import javax.ws.rs.core.Response;
 import javax.ws.rs.core.Response.Status;
 
@@ -12,12 +12,13 @@ import com.wakamayu.jucu.health.check.injector.enums.TypeStatus;
 import com.wakamayu.jucu.health.check.injector.interfaces.FactoryHealthCheck;
 import com.wakamayu.jucu.health.check.injector.model.ResponseHealth;
 
+import jakarta.enterprise.context.RequestScoped;
+
 import javax.inject.Inject;
-import javax.inject.Singleton;
+
 
 @Path("health")
-//@Consumes("application/json")
-//@Produces("application/json")
+@RequestScoped	
 public class HealthResource {
 
 	@Inject
